+++ src/main/java/com/cardiogenerator/generators/ECGDataGenerator.java
@@ -1,8 +1,8 @@
-package com.cardio_generator.generators;
+package com.cardiogenerator.generators;
 
 import java.util.Random;
 
-import com.cardio_generator.outputs.OutputStrategy;
+import com.cardiogenerator.outputs.OutputStrategy;
 
 public class ECGDataGenerator implements PatientDataGenerator {
     private static final Random random = new Random();
