+++ src/main/java/com/cardiogenerator/outputs/WebSocketOutputStrategy.java
@@ -1,4 +1,4 @@
-package com.cardio_generator.outputs;
+package com.cardiogenerator.outputs;
 
 import org.java_websocket.WebSocket;
 import org.java_websocket.server.WebSocketServer;
