+++ src/main/java/com/datamanagement/PatientRecord.java
@@ -1,4 +1,4 @@
-package com.data_management;
+package com.datamanagement;
 
 /**
  * Represents a single record of patient data at a specific point in time.
