+++ src/test/java/data_management/DataStorageTest.java
@@ -2,11 +2,11 @@ package data_management;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.data_management.DataReader;
 import org.junit.jupiter.api.Test;
 
-import com.data_management.DataStorage;
-import com.data_management.PatientRecord;
+import com.datamanagement.DataReader;
+import com.datamanagement.DataStorage;
+import com.datamanagement.PatientRecord;
 
 import java.util.List;
